requirements-p3.txt
cycler==0.10.0
fleep==1.0.1
joblib==0.13.2
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
opencv-python==4.0.1.24
pandas==0.24.2
pathlib==1.0.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.1
PyWavelets==1.0.3
scikit-learn==0.20.3
scipy==1.5.4
six==1.12.0
sklearn==0.0
tabulate==0.8.3
xgboost==1.4.2
numpy==1.19.5
